day4/pyproject.toml
#configure toml
[project]
name = "Banking"
version = "1.0.0"
description = "Banking Application"

readme = "README.md"
authors = [
    {name="Parameswari",email="Parameswaribala@gmail.com"}
]
requires-python = ">=3.8"
# production dependencies
dependencies = [
    "faker>=13.3.4",
    "pandas>=2.0.0",
    "python-dotenv>=0.19.0",
    "pydantic>=2.0.0"
   
]

#development dependencies
[project.optional-dependencies]
dev = [
    "pytest>=7.2.0",
    "pytest-html>=4.2.0",
    "pytest-cov>=3.0.0",
    "black>=23.3.0",
    "isort>=5.10.1",
    "sphinx>=5.0.0"
   
]

[tool.black]
line-length = 88
target-version = ['py313']
